feat: reconnect TCP when remote console loses connectivity
we already support when we lose our network connection, but
we don't disconnect when the remote console is removed from
the network, either intentionally or accidentally.

This updates the sending methods to check if we have a valid
TCP connection and aborts the connection if we cannot continue
to be connected.